Formatting FilterSearch.prod.jsx
